Adds get-auth and optional argument to get-user.
The get-auth function may be used to easily access the current
authentication map of friend.  This authentication map holds everything
related to the user from the mongo store besides the password hash.

The second, optional argument added to get-user (fields) allows for
selecting a subset of keys avaialble from a full query.