Modifies the way phrases are grabbed from mongo.
This modifies the get-phrase function, removing the dev-mode? check and
by correcting the use of the "random_point" field by querying with
[(rand) 0] instead of just (rand).

This also adds the add-trained-user-to-phrase function to add a user's
id to the users array for a phrase once they have completed training. As
well as adding this function. This commit adds a single call to it under
the user-profile route.